Centralize the Export/ExportDialog components in preparation for #113
